pyproject.toml
[tool.poetry]
name = "nekro-plugin-sec-run"
version = "1.1.0-dev"
description = "NekroAgent 安全Shell执行插件 - 执行安全的shell命令并返回图片"
authors = ["greenhandzdl"]
readme = "README.md"
packages = []
repository = "https://github.com/greenhandzdl/nekro-plugin-sec-run"
package-mode = false


[tool.poetry.dependencies]
python = ">=3.9,<3.12"
nekro-agent = ">=1.0.0"
Pillow = "^10.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

# [[tool.poetry.source]]
# name = "aliyun"
# url = "https://mirrors.aliyun.com/pypi/simple/"
# priority = "primary"

[tool.black]
line-length = 128

[tool.ruff]
extend-exclude = []
ignore = [
    "E402",
    "E501",
    "E722",
    "B008",
    "B905",
    "F401",
    "TID252",
    "TRY002",
    "TRY003",
    "RUF001",
    "RUF002",
    "RUF003",
    "RUF006",
    "RUF012",
    "RUF100",
    "PGH003",
    "N801",
    "A003",
]
select = [
    "F",
    "E",
    "I",
    "B",
    "A",
    "COM",
    "C4",
    "ISC",
    "PIE",
    "PYI",
    "Q",
    "RSE",
    "RET",
    "SLF",
    "SIM",
    "TID",
    "ARG",
    "PTH",
    "PGH",
    "TRY",
    "RUF",
]
